separate out etl processes
