+++ src/tasks/etl/__init__.py
@@ -1,0 +1,5 @@
+from auth_etl import AuthETL  # noqa: F401
+from single_activity_etl import SingleActivityETL  # noqa: F401
+from user_etl import UserETL  # noqa: F401
+from historic_activities_etl import ActivitiesETL  # noqa: F401
+from naming_etl import NameSuggestionETL  # noqa: F401
